+++ src/interfaces/energy-cli-conf.interface.ts
@@ -3,12 +3,18 @@ import { ItemTypes } from '../enums/item-types.enum';
 export interface IEnergyCLIConf {
     srcFolder: string;
     defaultExt: string;
-    additionalTypes: IEnergyAdditionalType[];
     dotnetInterfaceStyle: boolean;
+    additionalTypes: IEnergyAdditionalType[];
+    customFileTemplates: IEnergyCustomFileTemplate[];
 }
 
 export interface IEnergyAdditionalType {
     name: string;
     plural?: string;
     itemType?: ItemTypes;
+}
+
+export interface IEnergyCustomFileTemplate {
+    itemName: string;
+    template: string;
 }