+++ changelog.md
@@ -5,6 +5,7 @@
 * Added badges
 * Updated package repo data
 * Updated json schema (added customFileTemplates)
+* Updated json schema TS types
 
 ## 0.2.4 - Aelene
 * Throwing errors for: MissingItemTypeFlagException and unhandled errors.